+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * use CircleCi to run specs.
 * allow groups to have different semver level restrictions.
 * add ability to promote gems in the order of being updated
+* add ability to demote gems in the order of being updated
 
 ## 0.0.1 (2018-08-01)